refactor: #289 move to own package
