+++ src/main/java/biblivre/cataloging/holding/HoldingBO.java
@@ -30,12 +30,12 @@ import biblivre.cataloging.search.SearchDTO;
 import biblivre.circulation.user.UserBO;
 import biblivre.circulation.user.UserDTO;
 import biblivre.core.DTOCollection;
-import biblivre.core.LabelPrintDTO;
 import biblivre.core.configurations.ConfigurationBO;
 import biblivre.core.exceptions.ValidationException;
 import biblivre.core.file.DiskFile;
 import biblivre.core.utils.Constants;
 import biblivre.core.utils.ParagraphAlignmentUtil;
+import biblivre.labels.print.LabelPrintDTO;
 import biblivre.login.LoginBO;
 import biblivre.login.LoginDTO;
 import biblivre.marc.MarcDataReader;
