+++ src/main/java/biblivre/cataloging/labels/Handler.java
@@ -28,9 +28,9 @@ import biblivre.cataloging.holding.HoldingDTO;
 import biblivre.core.AbstractHandler;
 import biblivre.core.ExtendedRequest;
 import biblivre.core.ExtendedResponse;
-import biblivre.core.LabelPrintDTO;
 import biblivre.core.enums.ActionResult;
 import biblivre.core.file.DiskFile;
+import biblivre.labels.print.LabelPrintDTO;
 import biblivre.marc.MarcDataReader;
 import biblivre.marc.MarcUtils;
 import java.util.ArrayList;
