+++ src/main/java/biblivre/cataloging/labels/LabelGenerator.java
@@ -1,7 +1,7 @@
 package biblivre.cataloging.labels;
 
-import biblivre.core.LabelPrintDTO;
 import biblivre.core.file.DiskFile;
+import biblivre.labels.print.LabelPrintDTO;
 import java.util.List;
 
 public interface LabelGenerator {
