+++ src/main/java/biblivre/circulation/user/UserBO.java
@@ -23,10 +23,10 @@ import biblivre.administration.usertype.UserTypeBO;
 import biblivre.administration.usertype.UserTypeDTO;
 import biblivre.core.AbstractBO;
 import biblivre.core.DTOCollection;
-import biblivre.core.LabelPrintDTO;
 import biblivre.core.file.DiskFile;
 import biblivre.core.translations.TranslationsMap;
 import biblivre.core.utils.Constants;
+import biblivre.labels.print.LabelPrintDTO;
 import com.lowagie.text.Chunk;
 import com.lowagie.text.Document;
 import com.lowagie.text.Element;
