+++ src/main/java/biblivre/circulation/user_cards/Handler.java
@@ -23,9 +23,9 @@ import biblivre.circulation.user.UserBO;
 import biblivre.core.AbstractHandler;
 import biblivre.core.ExtendedRequest;
 import biblivre.core.ExtendedResponse;
-import biblivre.core.LabelPrintDTO;
 import biblivre.core.enums.ActionResult;
 import biblivre.core.file.DiskFile;
+import biblivre.labels.print.LabelPrintDTO;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.UUID;
